requirements.txt
absl-py==0.7.0
astor==0.7.1
cozmo==1.4.10
cozmoclad==3.4.0
cycler==0.10.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.3
mock==2.0.0
numpy==1.16.1
opencv-python==4.0.0.21
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
pyparsing==2.4.0
python-dateutil==2.8.0
six==1.12.0
tensorboard==1.13.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.31.1
Werkzeug==0.14.1
